See #50631 #50317. Class rules to OO API tests
1. New OO API implementations are tested with GDS types FBOONATIVE, FBOOEMBEDDED, FBOOLOCAL only;
2. Added separate jobs for testing with FBOONATIVE and FBOOEMBEDDED modes.

+++ ci/test.sh
@@ -174,7 +174,7 @@ else
 fi
 
 # Delete symlink for fbclient to test fbclient.jar
-if [[ "$GDS_TYPE" != "EMBEDDED" ]]; then
+if [[ "$GDS_TYPE" != "EMBEDDED" && "$GDS_TYPE" != "FBOOEMBEDDED" ]]; then
   rm -f /usr/lib64/libfbclient.so
 fi
